code: move types to api file
