+++ frontend/src/controller/api.ts
@@ -9,26 +9,15 @@ import type {
 } from '#api/types'
 import { FollowErrors, ProviderSignInErrors, ProviderSignUpErrors } from '#api/types'
 import { useAuthStore } from '@/stores/authStore'
-import { ApiErrorsGeneral, type ApiErrors } from '@/types'
+import {
+  ApiErrorsGeneral,
+  ResponseStatus,
+  type ApiErrors,
+  type ApiResponse,
+  type DetailedApiResponse,
+  type ResponseErrors
+} from '@/types'
 import ky, { HTTPError } from 'ky'
-
-export enum ResponseStatus {
-  OK = 200,
-  UNAUTHORIZED = 401,
-  BAD_REQUEST = 400,
-  SERVER_ERROR = 500
-}
-
-export type ResponseErrors = ResponseStatus.UNAUTHORIZED | ResponseStatus.BAD_REQUEST | ResponseStatus.SERVER_ERROR
-
-export type ResponseSuccess = ResponseStatus.OK
-
-export interface ApiResponse<T, S extends ResponseStatus = ResponseStatus> {
-  data: T
-  status: S
-}
-
-export type DetailedApiResponse<T> = ApiResponse<ApiErrors, ResponseErrors> | ApiResponse<T, ResponseSuccess>
 
 /**
  * ApiClient is a class that handles all requests to the api and handles non 200 responses
